edit readme and complete security horario

+++ src/controllers/horario.controller.ts
@@ -17,7 +17,7 @@ export class HorarioController{
             if (!token) {
                 return this.httpResponse.Unauthorized(res,"Inicie sesion primero");
             }
-            
+
             const data = await this.horarioService.findAll();
             if (data.length === 0) {
                 return this.httpResponse.NotFound(res, "No existen datos")
@@ -31,11 +31,21 @@ export class HorarioController{
     async getHorarioById(req: Request, res: Response) {
         const {id}= req.params;
         try {
-            const data = await this.horarioService.findbyid(id);
-            if (!data) {
-                return this.httpResponse.NotFound(res, "No existe datos")
+            // Verificamos si esta logeado o no 
+            const token = req.cookies.accessToken
+            if (!token) {
+                return this.httpResponse.Unauthorized(res,"Inicie sesion primero");
             }
-            return this.httpResponse.Ok(res, data)
+
+            const typeUser = await this.auth.setPermissions(token);
+            if (typeUser === 1 || typeUser === 2) {
+                const data = await this.horarioService.findbyid(id);
+                if (!data) {
+                    return this.httpResponse.NotFound(res, "No existe datos")
+                }
+                return this.httpResponse.Ok(res, data)
+            }
+            return this.httpResponse.Unauthorized(res,"No tiene permiso para esta accion");
         } catch (e) {
             return this.httpResponse.NotFound(res, e)
         }
@@ -43,8 +53,19 @@ export class HorarioController{
 
     async createHorario(req: Request, res: Response) {
         try {
-            const data = await this.horarioService.create(req.body);
-            return this.httpResponse.Ok(res, data)
+            // Verificamos si esta logeado o no 
+            const token = req.cookies.accessToken
+            if (!token) {
+                return this.httpResponse.Unauthorized(res,"Inicie sesion primero");
+            }
+
+            const typeUser = await this.auth.setPermissions(token);
+            if (typeUser === 1 || typeUser === 2) {
+
+                const data = await this.horarioService.create(req.body);
+                return this.httpResponse.Ok(res, data)
+            }
+            return this.httpResponse.Unauthorized(res,"No tiene permiso para esta accion");
         } catch (e) {
             return this.httpResponse.NotFound(res, e)
         }
@@ -53,13 +74,24 @@ export class HorarioController{
     async updateHorario(req: Request, res: Response) {
         const {id}= req.params;
         try {
-            const data = await this.horarioService.update(id,req.body);
-            
-            if (!data.affected) {
-                return this.httpResponse.NotFound(res, "Hay un error al actualizar")
+            // Verificamos si esta logeado o no 
+            const token = req.cookies.accessToken
+            if (!token) {
+                return this.httpResponse.Unauthorized(res,"Inicie sesion primero");
             }
 
-            return this.httpResponse.Ok(res, data)
+            const typeUser = await this.auth.setPermissions(token);
+            if (typeUser === 1 || typeUser === 2) {
+                const data = await this.horarioService.update(id,req.body);
+                
+                if (!data.affected) {
+                    return this.httpResponse.NotFound(res, "Hay un error al actualizar")
+                }
+    
+                return this.httpResponse.Ok(res, data)
+            }
+            
+            return this.httpResponse.Unauthorized(res,"No tiene permiso para esta accion");
         } catch (e) {
             return this.httpResponse.NotFound(res, e)
         }
@@ -68,13 +100,24 @@ export class HorarioController{
     async deteleHorario(req: Request, res: Response) {
         const {id}= req.params;
         try {
-            const data = await this.horarioService.delete(id);
-            
-            if (!data.affected) {
-                return this.httpResponse.NotFound(res, "Hay un error al actualizar")
+            // Verificamos si esta logeado o no 
+            const token = req.cookies.accessToken
+            if (!token) {
+                return this.httpResponse.Unauthorized(res,"Inicie sesion primero");
             }
 
-            return this.httpResponse.Ok(res, data)
+            const typeUser = await this.auth.setPermissions(token);
+            if (typeUser === 1 ) {
+                const data = await this.horarioService.delete(id);
+                
+                if (!data.affected) {
+                    return this.httpResponse.NotFound(res, "Hay un error al eliminar")
+                }
+    
+                return this.httpResponse.Ok(res, data)
+            }
+            
+            return this.httpResponse.Unauthorized(res,"No tiene permiso para esta accion");
         } catch (e) {
             return this.httpResponse.NotFound(res, e)
         }
